Refaktorering for mer oversiktlig kode.
